update(test): reorganizing pom for videocalls
